+++ CHANGES.md
@@ -10,6 +10,7 @@
    * Array schemas can now specify a `minCount` and a `maxCount`.
    * Schemas can now be specified using both JSON or SNBT.
    * Object schemas with a `"default-property"` can now also specify a `"default-key"` schema.
+   * Added suggestions for arguments of templates in schema definitions. 
 
 ### Fixes
  * Fixed crash when attempting to load project with a datapack version that is not supported.